API

API

RHEOS Types

RHEOS.RheologyData — Type.

RheologyData(σ::Vector{T}, ϵ::Vector{T}, t::Vector{T}, sampling::String, log::Vector{String}) where T<:Real

RheologyData struct contains stress, strain and time data.

If preferred, an instance can be generated manually by just providing the three data vectors in the right order, sampling type will be checked automatically. If loading partial data (either stress or strain), fill the other vector as a vector of zeros of the same length as the others.

Fields

σ: stress

ϵ: strain

t: time

sampling: sampling type, either "constant" or "variable"

log: a log of struct's events, e.g. preprocessing

source

RHEOS.RheologyDynamic — Type.

RheologyDynamic(Gp::Vector{T}, Gpp::Vector{T}, ω::Vector{T}, log::Vector{String}) where T<:Real

RheologyDynamic contains storage modulus, loss modulus and frequency data.

If preferred, an instance can be generated manually by just providing the three data vectors in the right order.

Fields

Gp: storage modulus

Gpp: loss modulus

ω: frequency

log: a log of struct's events, e.g. preprocessing

source

RHEOS.RheologyModel — Type.

RheologyModel(G::T, J::T, Gp::T, Gpp::T, parameters::Vector{S<:Real} log::Vector{String}) where T<:Function

RheologyModel contains a model's various moduli, parameters, and log of activity.

For incomplete models, an alternative constructor is available where all arguments are keyword arguments and moduli not provided default to a null modulus which always returns [-1.0].

Fields

G: Relaxation modulus

J: Creep modulus

Gp: Storage modulus

Gpp: Loss modulus

parameters: Used for predicting and as default starting parameters in fitting

log: a log of struct's events, e.g. what file it was fitted to

source

Preprocessing Functions