Save and Load

RHEOS.savedata — Method.

savedata(self::RheologyData; filedir::String = "", ext = "_RheologyData.jld")

Save RheologyData object using JLD format. Save file directory must be specified. If data was loaded from disk using fileload or the full RheologyData constructor then filedir argument can be set to empty string "" which will try to use the original file dir concatenated with the optional ext string argument - e.g. "/originalpathto/file.csv_RheologyData.jld".

source

RHEOS.loaddata — Method.

loaddata(filedir::String)

Convenience function loads RheologyData.

source

RHEOS.savemodel — Method.

savemodel(self::RheologyModel; filedir::String = "", ext = "")

source

RHEOS.loadmodel — Method.

loadmodel(filedir::String)

Convenience function loads RheologyModel from disk.

source

RHEOS.exportdata — Method.

exportdata(self::RheologyData; filedir::String = "", ext = "_mod.csv")

Export RheologyData to csv format. Exports three columns in order: stress, strain, time. Useful for plotting/analysis in other software.

source