Models

Models

Models included in RHEOS

Severa models (i.e. their relaxation modulus, creep modulus and complex modulus) already implemented in RHEOS. Their constructors are listed below. Note that the params argument is always optional, as represented by their containment in square barckets. If left blank then the model's default parameters are used. Additional models are forthcoming.

Elements

RHEOS.Spring — Function.

Spring([params::Vector{T}]) where T<:Real

Single spring element.

source

RHEOS.DashPot — Function.

DashPot([params::Vector{T}]) where T<:Real

Single dashpot element.

source

RHEOS.SpringPot — Function.

SpringPot([params::Vector{T}]) where T<:Real

Single springpot element.

source

Maxwell Type

RHEOS.FractionalMaxwell — Function.

FractionalMaxwell([params::Vector{T}]) where T<:Real

Two springpots in series.

source

RHEOS.FractionalMaxwellSpring — Function.

FractionalMaxwellSpring([params::Vector{T}]) where T<:Real

A springpot and spring in series.

source

RHEOS.FractionalMaxwellDashpot — Function.

FractionalMaxwellDashpot([params::Vector{T}]) where T<:Real

A springpot and dashpot in series.

source

RHEOS.Maxwell — Function.

Maxwell([params::Vector{T}]) where T<:Real

A spring and dashpot in series.

source

Kelvin-Voigt Type

RHEOS.FractionalKelvinVoigt — Function.

FractionalKelvinVoigt([params::Vector{T}]) where T<:Real

Two springpots in parallel.

source

RHEOS.FractionalKVspring — Function.

FractionalKVspring([params::Vector{T}]) where T<:Real

A springpot and spring in parallel.

source

RHEOS.FractionalKVdashpot — Function.

FractionalKVdashpot([params::Vector{T}]) where T<:Real

A springpot and dashpot in parallel.

source

RHEOS.KelvinVoigt — Function.

KelvinVoigt([params::Vector{T}]) where T<:Real

A spring and dashpot in parallel.

source

Zener Type

RHEOS.FractionalZener — Function.

FractionalZener([params::Vector{T}]) where T<:Real

2 springpots in series, both in parallel with a springpot.

source

RHEOS.FractionalSLS — Function.

FractionalSLS([params::Vector{T}]) where T<:Real

A springpot and spring in series, both in parallel with a spring.

source

RHEOS.FractionalSpecial — Function.

FractionalSpecial([params::Vector{T}]) where T<:Real

A springpot and dashpot in series, both in parallel with a spring.

source

RHEOS.SLS — Function.

SLS(params::Vector{T}) where T<:Real

A spring and dashpot in series, both in parallel with a spring.

source

Creating your own model

If you know some (or all) of the moduli for a model that you would like use but hasn't already been implemented in RHEOS, this section will explain how to quickly import these moduli into a RheologyModel object for use with other parts of RHEOS. For the sake of example, we will use the relaxation modulus, storage modulus and loss modulus of the Standard Linear Solid model as defined in RHEOS.

function G_sls(t::Vector{T}, params::Vector{T}) where T<:Real
    η, kᵦ, kᵧ = params

    G = kᵧ .+ kᵦ*exp.(-t*kᵦ/η)
end

function Gp_sls(ω::Vector{T}, params::Vector{T}) where T<:Real
    η, kᵦ, kᵧ = params

    τ = η/kᵦ

    denominator = 1 .+ (τ^2)*(ω.^2)
    numerator = (ω.^2)*(τ^2)*kᵦ

    Gp = numerator./denominator .+ kᵧ
end

function Gpp_sls(ω::{T}, params::Vector{T}) where T<:Real
    η, kᵦ, kᵧ = params

    τ = η/kᵦ

    denominator = 1 .+ (τ^2)*(ω.^2)
    numerator = ω*τ*kᵦ

    Gpp = numerator./denominator
end

Now we have the our moduli defined as Julia functions we can store them, along with some (optional) default parameters, in a RheologyModel struct in the following way.

our_model = RheologyModel(G = G_sls, Gp = Gp_sls, Gpp = Gpp_sls, params = [1.0, 0.5, 1.0])

Now we can fit this model to data and use it to make predictions. Any moduli not included in this final step will default to a null_modulus which always returns the array [-1.0].