Preprocessing

RHEOS.var_resample — Method.

var_resample(self::RheologyData, refvar::Symbol, pcntdownsample::Float64; mapback::Bool = false)

Convert a fixed sample rate array to a variable sample rate, with sampling points added according to a relative change in chosen variable refvar, 1st derivative of refvar and 2nd derivative of refvar (WRT time). Usually chosen as the measured variable, so :σ for a stress relaxation test and :ϵ for a creep test.

Currently only variable downsampling supported. pcntdown sample is approximate, works well in some cases and very poorly in others. If required, compare resampled length vs original length after processing has finished. If data is noisy, may benefit from sending smoothed signal to this algorithm and either using mapback function or interpolating onto unsmoothed data.

See help docstring for var_resample for more details on algorithm implementation.

source

RHEOS.downsample — Method.

downsample(self::RheologyData, time_boundaries::Array{Float64,1}, elperiods::Array{Int64,1})

High-level RheologyData interface to downsample in base.jl. Boundaries are floating point times which are then converted to the closest elements.

source

RHEOS.fixed_resample — Method.

fixed_resample(self::RheologyData, time_boundaries::Array{Float64,1}, elperiods::Array{Int64,1}, direction::Array{String,1})

High-level RheologyData interface to fixed_resample in base.jl

source

RHEOS.smooth — Method.

smooth(self::RheologyData, τ::Float64)

Smooth data using a Gaussian Kernel to time scale τ (approximately half power).

Smooths both σ and ϵ.

source

function mapbackdata(self_new::RheologyData, self_original::RheologyData)