API

API

This page is a list of all types and functions defined by RHEOS that are intended for general use.

RHEOS Types

RHEOS.RheologyData — Type.

RheologyData(σ::Vector{T}, ϵ::Vector{T}, t::Vector{T}, sampling::String, log::Vector{String}) where T<:Real

RheologyData struct contains stress, strain and time data.

If preferred, an instance can be generated manually by just providing the three data vectors in the right order, sampling type will be checked automatically. If loading partial data (either stress or strain), fill the other vector as a vector of zeros of the same length as the others.

Fields

σ: stress

ϵ: strain

t: time

sampling: sampling type, either "constant" or "variable"

log: a log of struct's events, e.g. preprocessing

source

RHEOS.RheologyDynamic — Type.

RheologyDynamic(Gp::Vector{T}, Gpp::Vector{T}, ω::Vector{T}, log::Vector{String}) where T<:Real

RheologyDynamic contains storage modulus, loss modulus and frequency data.

If preferred, an instance can be generated manually by just providing the three data vectors in the right order.

Fields

Gp: storage modulus

Gpp: loss modulus

ω: frequency

log: a log of struct's events, e.g. preprocessing

source

RHEOS.RheologyModel — Type.

RheologyModel(G::T, J::T, Gp::T, Gpp::T, parameters::Vector{S<:Real} log::Vector{String}) where T<:Function

RheologyModel contains a model's various moduli, parameters, and log of activity.

For incomplete models, an alternative constructor is available where all arguments are keyword arguments and moduli not provided default to a null modulus which always returns [-1.0].

Fields

G: Relaxation modulus

J: Creep modulus

Gp: Storage modulus

Gpp: Loss modulus

parameters: Used for predicting and as default starting parameters in fitting

log: a log of struct's events, e.g. what file it was fitted to

source

Preprocessing Functions

RHEOS.downsample — Function.

downsample(self::RheologyData, time_boundaries::Vector{T} where T<:Real, elperiods::Vector{S} where S<:Integer)

Boundaries are floating point times which are then converted to the closest elements. Works by just reducing on indices. For example, time_boundaries could be [0.0, 10.0, 100.0] and elperiods could be [2, 4]. So new data would take every 2nd element from 0.0 seconds to 10.0 seconds, then every 4th element from 10.0 seconds to 100.0 seconds.

source

RHEOS.fixedresample — Function.

fixedresample(self::RheologyData, time_boundaries::Vector{T} where T<:Real, elperiods::Vector{K} where K<:Integer, direction::Vector{String})

Resample data with new sample rate(s).

Whereas downsample can only reduce the sample rate by not taking every array element, fixedresample can also upsample. Whether to up or down sample for a given section is known from the direction argument.

source

RHEOS.variableresample — Function.

variableresample(self::RheologyData, refvar::Symbol, pcntdownsample::Real; mapback::Bool = false)

Convert a fixed sample rate array to a variable sample rate, with sampling points added according to a relative change in chosen variable refvar, 1st derivative of refvar and 2nd derivative of refvar (WRT time). Usually chosen as the measured variable, so :σ for a stress relaxation test and :ϵ for a creep test.

Currently only variable downsampling supported. pcntdown sample is approximate, works well in some cases and very poorly in others. If required, compare resampled length vs original length after processing has finished. If data is noisy, may benefit from sending smoothed signal to this algorithm and either using mapback function or interpolating onto unsmoothed data.

Algorithm works as follows. 25 initial samples are generated evenly spread. After this, the array is repeatedly sweeped, anywhere Δy, Δdy/dx, Δd2y/dx2 is greater than a threshold, α, a new sample is created at the midpoint of the two tested points. This is allowed to happen a maximum of 400 times, after which α is decreased and the process starts again. This macro process continues until the desired pcntdownsample ratio has been reached.

Arguments

self: RheologyData instance

refvar: The data whose derivatives will determine sample densities

pcntdownsample: Approximate ratio of new samples to old samples

_mapback = false: (Optional) Determines whether resampled points should 'snap' to closest original points

source

RHEOS.smooth — Function.

smooth(self::RheologyData, τ::Real; pad::String="replicate")

Smooth data using a Gaussian Kernel to time scale τ (approximately half power).

Smooths both σ and ϵ. Essentially a low pass filter with frequencies of 1/τ being cut to approximately half power. For other pad types available see ImageFiltering documentation.

source

RHEOS.zerotime — Function.

function zerotime(self::RheologyData)

Convenience function to normalize time such that the starting time is 0.0

source

Fitting and Predicting Functions

RHEOS.modelfit — Function.

modelfit(data::RheologyData, model::RheologyModel, modtouse::Symbol; p0::Vector{T} = [-1.0], lo::Vector{T} = [-1.0], hi::Vector{T} = [-1.0], verbose::Bool = false, rel_tol = 1e-4, diff_method="BD") where T<:Real

Fit RheologyData struct to model and return a fitted model as a RheologyModel object.

Arguments

data: RheologyData struct containing all data

model: RheologyModel containing moduli and default (initial) parameters

modtouse: :G for relaxation modulus, :J for creep modulus

p0: Initial parameters to use in fit (uses 'model' parameters if none given)

lo: Lower bounds for parameters

hi: Upper bounds for parameters

verbose: If true, prints parameters on each optimisation iteration

rel_tol: Relative tolerance of optimization, see NLOpt docs for more details

diff_method: Set finite difference formula to use for derivative, currently "BD" or "CD"

source

RHEOS.modelpredict — Function.

modelpredict(data::RheologyData, model::RheologyModel, modtouse::Symbol; diff_method="BD")

Given data and model, return new dataset based on model parameters and using the modulus specified by 'modtouse'; either creep modulus (:J, only returned strain is new) or relaxation modulus (:G, only returned stress is new). 'diff_method' sets finite difference for calculating the derivative used in the hereditary integral and can be either backwards difference ("BD") or central difference ("CD").

source

RHEOS.modelstepfit — Function.

modelstepfit(data::RheologyData, model::RheologyModel, modtouse::Symbol; p0::Vector{T} = [-1.0], lo::Vector{T} = [-1.0], hi::Vector{T} = [-1.0], verbose::Bool = false, rel_tol = 1e-4) where T<:Real

Same as 'modelfit' except assumes a step loading. If this assumption is appropriate for the data then fitting can be sped up greatly by use of this function. If modtouse is :G, relaxation modulus, then the first element of the strain is assumed to be the amplitude of the step. If modtouse is :j, creep modulus, then the first element of the stress is assumed to be the amplitude of the step.

Arguments

data: RheologyData struct containing all data

model: RheologyModel containing moduli and default (initial) parameters

modtouse: :G for relaxation modulus, :J for creep modulus

p0: Initial parameters to use in fit (uses 'model' parameters if none given)

lo: Lower bounds for parameters

hi: Upper bounds for parameters

verbose: If true, prints parameters on each optimisation iteration

rel_tol: Relative tolerance of optimization, see NLOpt docs for more details

source

RHEOS.modelsteppredict — Function.

modelsteppredict(data::RheologyData, model::RheologyModel, modtouse::Symbol; step_on::Real = 0.0)

Same as modelpredict but assumes a step loading with step starting at 'step_on'. Singularities are bypassed by adding 1 to the index of the singular element.

source

RHEOS.dynamicmodelfit — Function.

dynamicmodelfit(data::RheologyDynamic, model::RheologyModel; p0::Vector{T} = [-1.0], lo::Vector{T} = [-1.0], hi::Vector{T} = [-1.0], verbose::Bool = false, rel_tol = 1e-4) where T<:Real

Fits model to the frequency/loss+storage moduli data.

All arguments are as described below. The 'weights' argument some more information. As this fitting procedure is fitting two functions simultaneously (the storage and loss moduli), if left untransformed the fit would tend to favour the modulus which is larger in magnitude and not fit the other modulus well. To avoid this, RHEOS offers a number of data transforms which can be used.

Arguments

data: RheologyDynamic struct containing all data

model: RheologyModel containing moduli and default (initial) parameters

p0: Initial parameters to use in fit (uses 'model' parameters if none given)

lo: Lower bounds for parameters

hi: Upper bounds for parameters

verbose: If true, prints parameters on each optimisation iteration

rel_tol: Relative tolerance of optimization, see NLOpt docs for more details

weights: Weighting mode for storage and loss modulus (see above)

source

RHEOS.dynamicmodelpredict — Function.

dynamicmodelpredict(data::RheologyDynamic, model::RheologyModel)

Given dynamic rheology data and model, return new dataset based on model parameters. Returns another RheologyDynamic instance with the predicted Gp and Gpp based on the frequencies and model parameters.

source

Data Generation Functions

RHEOS.linegen — Function.

linegen(t_total::Real; stepsize::Real = 1.0)

Generate RheologyData struct with a simple line loading of height 1.0.

Arguments

t_total: Total time length of data

stepsize: Time sampling period

source

RHEOS.stepgen — Function.

stepgen(t_total::Real, t_on::Real; t_trans::Real = 0.0, stepsize::Real = 1.0)

Generate RheologyData struct with a step loading of height 1.0. If t_trans is 0.0 then the step is instantaneous, otherwise the step is approximated by a logistic function approximately centered at t_on.

Arguments

t_total: Total time length of data

t_on: Step on time

t_trans: Step transition time

stepsize: Time sampling period

source

RHEOS.rampgen — Function.

rampgen(t_total::Real, t_start::Real, t_stop::Real; amplitude::Real = 1.0, baseval::Real = 0.0, stepsize::Real = 1.0)

Generate RheologyData struct with a ramp function. Reaches amplitude of 1.0 at t_stop.

Arguments

t_total: Total time length of data

t_start: Time for starting ramp

t_stop: Time of stopping ramp

stepsize: Time sampling period

source

RHEOS.singen — Function.

singen(t_total::Real, frequency::Real; t_start::Real = 0.0, phase::Real = 0.0, stepsize::Real = 1.0)

Generate RheologyData struct with a sinusoidal loading of amplitude 1.0.

Arguments

t_total: Total time length of data

frequency: Frequency of oscillation (Hz)

t_start: Time for oscillation to begin

phase: Phase of oscillation (radians)

stepsize: Time sampling period

source

RHEOS.noisegen — Function.

noisegen(t_total::Real; seed::Union{Int, Nothing} = nothing, stepsize::Real = 1.0)

Generate uniform random noise of maximum amplitude +/- 1.0. If reproducibility is required, always use the same number in the seed keyword argument with the same non-negative integer.

Arguments

t_total: Total time length of data

seed: Seed used for random number generation

baseval: Initial amplitude before oscillation started

stepsize: Time sampling period

source

RHEOS.repeatdata — Function.

repeatdata(self::RheologyData, n::Integer)

Repeat a given RheologyData data set n times.

source

File I/O

RHEOS.importdata — Function.

importdata(colnames::Array{String,1}, filedir::String; delimiter=',')

Load data from a CSV file (two/three columns, comma seperated by default but delimiter can be specified in the delimiter keyword argument). Columns must be identified by providing an array of strings which tell the function which data is contained in each column.

Can be used to construct either a RheologyData instance or a RheologyDynamic instance. Function detects whether "time" or "frequency" has been included and proceeds accordingly. For oscillatory data, all three columns (Gp, Gpp, Frequency) must be provided. For regular viscoelastic data two or three columns can be provided but time must always be provided. I.e. either stress/strain/time, stress/time or strain/time.

source

RHEOS.exportdata — Function.

exportdata(self::Union{RheologyData, RheologyDynamic}, filedir::String; ext = ".csv", delimiter=',')

Export RheologyData or RheologyDynamic type to csv format. Exports three columns in order: stress, strain, time for standard viscoelastic data. Or: Gp, Gpp, frequency for oscillatory data. File extension can be modified using the ext keyword argument. As with importdata, the delimiter can also be set by keyword argument.

Useful for plotting/analysis in other software.

source

RHEOS.savedata — Function.

savedata(self::Union{RheologyData, RheologyDynamic}, filedir::String; ext = ".jld2")

Save RheologyData or RheologyDynamic object using JLD2 format reuse in a later Julia session.

source

RHEOS.loaddata — Function.

loaddata(filedir::String)

Loads RheologyData or RheologyDynamic object from a jld2 file.

source

RHEOS.savemodel — Function.

savemodel(self::RheologyModel, filedir::String; ext = ".jld2")

Save RheologyModel object using JLD2 format reuse in a later Julia session.

source

RHEOS.loadmodel — Function.

loaddata(filedir::String)

Loads RheologyModel from a JLD2 file.

source