API

Main RHEOS Structs

RHEOS.RheoTimeData — Type

RheoTimeData(;σ::Vector{T1}, ϵ::Vector{T2}, t::Vector{T3}) where {T1<:Real, T2<:Real, T3<:Real}

RheoTimeData struct contains stress, strain and time data.

If preferred, an instance can be generated manually by just providing the three data vectors in the right order, sampling type will be checked automatically.

Fields

σ: stress

ϵ: strain

t: time

log: a log of struct's events, e.g. preprocessing

RHEOS.RheoFreqData — Type

RheoFreqData(Gp::Vector{T1}, Gpp::Vector{T2}, ω::Vector{T3}, log::OrderedDict{Any,Any}) where {T1<:Real, T2<:Real, T3<:Real}

RheoFreqData contains storage modulus, loss modulus and frequency data.

If preferred, an instance can be generated manually by just providing the three data vectors in the right order.

Fields

Gp: storage modulus

Gpp: loss modulus

ω: frequency

log: a log of struct's events, e.g. preprocessing

RHEOS.RheoModelClass — Type

RheoModelClass(name::String, params::Vector{Symbol}, _G::FunctionWrapper{RheoFloat,Tuple{RheoFloat,Vector{RheoFloat}}}, _Ga::FunctionWrapper{Vector{RheoFloat},Tuple{Vector{RheoFloat},Vector{RheoFloat}}}, _J::FunctionWrapper{RheoFloat,Tuple{RheoFloat,Vector{RheoFloat}}}, _Ja::FunctionWrapper{Vector{RheoFloat},Tuple{Vector{RheoFloat},Vector{RheoFloat}}}, _Gp::FunctionWrapper{RheoFloat,Tuple{RheoFloat,Vector{RheoFloat}}}, _Gpa::FunctionWrapper{Vector{RheoFloat},Tuple{Vector{RheoFloat},Vector{RheoFloat}}}, _Gpp::FunctionWrapper{RheoFloat,Tuple{RheoFloat,Vector{RheoFloat}}}, _Gppa::FunctionWrapper{Vector{RheoFloat},Tuple{Vector{RheoFloat},Vector{RheoFloat}}}, constraint::FunctionWrapper{Bool,Tuple{Vector{RheoFloat}}}, info::String, expressions::NamedTuple)

RheoModelClass contains a model name, it's symbolic parameters and all its moduli (both single-input and array-input versions).

It also contains information about any constraints that must be observed (e.g. the springpot coefficient being inbetween 0 and 1).

Lastly, it also contains additional info about the model which may include a text-art schematic.

Generally, users will want to use the RheoModelClass constructor function as shown in the 'Create Your Model' section of the documentation rather than the default constructor.

RHEOS.RheoModel — Type

RheoModel(_G::FunctionWrapper{RheoFloat,Tuple{RheoFloat}}, _Ga::FunctionWrapper{Vector{RheoFloat},Tuple{Vector{RheoFloat}}}, _J::FunctionWrapper{RheoFloat,Tuple{RheoFloat}}, _Ja::FunctionWrapper{Vector{RheoFloat},Tuple{Vector{RheoFloat}}}, _Gp::FunctionWrapper{RheoFloat,Tuple{RheoFloat}}, _Gpa::FunctionWrapper{Vector{RheoFloat},Tuple{Vector{RheoFloat}}}, _Gpp::FunctionWrapper{RheoFloat,Tuple{RheoFloat}}, _Gppa::FunctionWrapper{Vector{RheoFloat},Tuple{Vector{RheoFloat}}}, expressions::NamedTuple)

RheoModel contains all known moduli of a particular model, as for a RheoModelClass model name. However, a RheoModel has all its parameters fixed to known values.

Generally, users will begin with a defined RheoModelClass (e.g. SLS) and then specialise the parameters, rather than calling the default constructor explicitly.

Moduli/compliance functions

RHEOS.relaxmod — Function

relaxmod(m[, t, params])

provide access to the relaxation modulus (G) of a given model m. It can be used with a broad range of inputs.

When a time value is provided (or an array of time values), the function returns the corresponding value(s) of the relaxation modulus.

relaxmod(m::RheoModel, time (single value or array))

relaxmod(m::RheoModelClass, time (single value or array), parameters (array, named tupple or keyword parameters))

Examples:

relaxmod(Maxwell, 1, k=1., η=1)

relaxmod(Maxwell, [1,2,3], [1,1])

When no time value is provided, relaxmod returns the relaxation function itself.

relaxmod(m::RheoModel)

relaxmod(m::RheoModelClass, parameters (array, named tupple or keyword parameters))

Examples:

m = RheoModel(Maxwell, k=1., η=1)
G = relaxmod(m)
G([0,1,2])

RHEOS.creepcomp — Function

creepcomp(m[, t, params])

provide access to the creep compliance (J) of a given model m. It can be used with a broad range of inputs.

When a time value is provided (or an array of time values), the function returns the corresponding value(s) of the creep compliance.

creepcomp(m::RheoModel, time (single value or array))

creepcomp(m::RheoModelClass, time (single value or array), parameters (array, named tupple or keyword parameters))

Examples:

creepcomp(Maxwell, 1, k=1., η=1)

creepcomp(Maxwell, [1,2,3], [1,1])

When no time value is provided, creepcomp returns the creep compliance function itself.

creepcomp(m::RheoModel)

creepcomp(m::RheoModelClass, parameters (array, named tupple or keyword parameters))

Examples:

m = RheoModel(Maxwell, k=1., η=1)
J = creepcomp(m)
J([0,1,2])

RHEOS.storagemod — Function

storagemod(m[, ω, params])

provide access to the storage modulus (G') of a given model m.

When a frequency value is provided (or an array of frequency values), the function returns the corresponding value(s) of the storage modulus.

storagemod(m::RheoModel, frequency (single value or array))

storagemod(m::RheoModelClass, frequency (single value or array), parameters (array, named tupple or keyword parameters))

Examples:

storagemod(Maxwell, 1, k=1., η=1)

storagemod(Maxwell, [1,2,3], [1,1])

When no time value is provided, storagemod returns the storage modulus function itself.

storagemod(m::RheoModel)

storagemod(m::RheoModelClass, parameters (array, named tupple or keyword parameters))

Examples:

m = RheoModel(Maxwell, k=1., η=1)
Gp = storagemod(m)
Gp([0,1,2])

RHEOS.lossmod — Function

lossmod(m[, ω, params])

provide access to the loss modulus (G'') of a given model m.

When a frequency value is provided (or an array of frequency values), the function returns the corresponding value(s) of the loss modulus.

lossmod(m::RheoModel, frequency (single value or array))

lossmod(m::RheoModelClass, frequency (single value or array), parameters (array, named tupple or keyword parameters))

Examples:

lossmod(Maxwell, 1, k=1., η=1)

lossmod(Maxwell, [1,2,3], [1,1])

When no time value is provided, lossmod returns the loss modulus function itself.

lossmod(m::RheoModel)

lossmod(m::RheoModelClass, parameters (array, named tupple or keyword parameters))

Examples:

m = RheoModel(Maxwell, k=1., η=1)
Gpp = lossmod(m)
Gpp([0,1,2])

RHEOS.dynamicmod — Function

dynamicmod(m[, ω, params])

provide access to the complex dynamic modulus (G' + i G'') of a given model m.

When a frequency value is provided (or an array of frequency values), the function returns the corresponding value(s) of the complex dynamic modulus.

dynamicmod(m::RheoModel, frequency (single value or array))

dynamicmod(m::RheoModelClass, frequency (single value or array), parameters (array, named tupple or keyword parameters))

Examples:

dynamicmod(Maxwell, 1, k=1., η=1)

dynamicmod(Maxwell, [1,2,3], [1,1])

When no time value is provided, lossmod returns the complex dynamic modulus function itself.

dynamicmod(m::RheoModel)

dynamicmod(m::RheoModelClass, parameters (array, named tupple or keyword parameters))

Examples:

m = RheoModel(Maxwell, k=1., η=1)
Gpp = dynamicmod(m)
Gpp([0,1,2])

Note: use abs() and angle() to get the magnitude and phase of the complex modulus.

Sampling and Filtering Functions

RHEOS.resample — Function

resample(d::RheoTimeData [, t = array of time value or range, dt = time step required, scale = multiplicator to apply to existing sampling rate] )

Resample data.

Usage:

'resample(d)' would keep the number of sampling points the same but interpolate to set a uniform time step.

'resample(d, t=-1:0.1:10)' would resample by interpolation to generate a new dataset with time points given by the range or array values passed with keyword 't'.

'resample(d, dt=0.1)' would resample by interpolation to generate a new dataset with time step 'dt'.

'resample(d, scale=2)' would resample by multiplying the sampling rate by 'scale'. This could down-sample (scale<1) or upsample (scale>1). If timesteps are non uniform, it would interpolate values accordingly.

RHEOS.indexweight — Function

indexweight(self::RheoTimeData, elperiods::Union{Vector{K}, K}; time_boundaries::Union{Nothing, Vector{T}} = nothing, includelast=true) where {K<:Integer,T<:Real}

This function returns array indices (i.e. an array of integers) which can be sent to the modelfit, modelstepfit or modeldiffeqfit functions to provide a weighted fitting whilst maintaining constant sample-rate.

indexweight can underweight indices or overweight them. If elperiods in a given boundary is negative, every abs(n) index will be used, where n is the elperiod corresponding to a given boundary. If n is positive, then indicies will be duplicated n times such that they are given a higher weighting during the fitting procedure. If time_boundaries are not specified, resampling is applied to the whole set of data. If number of elements per period (elperiods) is 1 or -1 it returns the original indicies for that boundary, whilst 0 is not accepted as a valid argument for elperiods.

The last element may or may not be included. By default the last element is forced to be included but this can be negated by providing the keyword argument includelast=false.

RHEOS.cutting — Function

cutting(self::RheoTimeData, time_on::T1, time_off::T2) where {T1<:Number, T2<:Number}

Remove the data outside a specified time interval.

By specifing a time interval (time_on, time_off), a new RheoTimeData is returned without the data lying outside time interval.

RHEOS.smooth — Function

smooth(self::RheoTimeData, τ::Real; pad::String="reflect")

Smooth data using a Gaussian Kernel to time scale τ (approximately half power).

Smooths both σ and ϵ. Sampling frequency must be constant as it is based on FFT convolution. Essentially a low pass filter with frequencies of 1/τ being cut to approximately half power. For other pad types available see ImageFiltering documentation. As of doc writing, pad options are: "replicate" (repeat edge values to infinity), "circular" (image edges "wrap around"), "symmetric" (the image reflects relative to a position between pixels), "reflect" (the image reflects relative to the edge itself).

RHEOS.extract — Function

extract(self::Union{RheoTimeData,RheoFreqData}, type::Union{TimeDataType,FreqDataType,Integer})

Extract specific fields form RheoTimeData or RheoFreqData.

Extract can copy one or more fields from a given RheoXData variable into a new RheoXData one. The fields that are copied are identified by the specified type of data. If self is a RheoTimeData, the type that can be extracted is time_only (or 0), stress_only (or 1), strain_only (or 2). Note that strain_and_stress (or 3) is not allowed. If self is a RheoFreqData, the type that can be extracted is freq_only (or 0).

Fitting and Predicting Functions

RHEOS.modelfit — Function

modelfit(data::RheoTimeData, model::RheoModelClass, modloading::Symbol; p0::Union{NamedTuple,Nothing} = nothing, lo::Union{NamedTuple,Nothing} = nothing, hi::Union{NamedTuple,Nothing} = nothing, verbose::Bool = false, rel_tol = 1e-4, diff_method="BD", weights::Union{Vector{Integer},Nothing}=nothing)

Fit RheologyData struct to model and return a fitted model as a RheologyModel object. For the fitting process RHEOS relies on the optimistion package NLopt.jl (https://nlopt.readthedocs.io/en/latest/). RHEOS makes use of a local derivative free algorithm, specifically the Tom Rowan's "Subplex"

Arguments

data: RheoTimeData struct containing all data

model: RheoModelClass containing moduli functions and named tuple parameters

modloading: strain_imposed or 1, stress_imposed or 2

p0: Initial parameters to use in fit (uses 0.5 for all parameters if not defined)

lo: Lower bounds for parameters

hi: Upper bounds for parameters

verbose: If true, prints parameters on each optimisation iteration

rel_tol: Relative tolerance of optimization, see NLOpt docs for more details

diff_method: Set finite difference formula to use for derivative, currently "BD" or "CD"

weights: Vector of indices weighted according to importance, can be generated by indexweight function

RHEOS.modelpredict — Function

modelpredict(data::RheoTimeData, model::RheoModel; diff_method="BD")

Given an incomplete data set (only either stress or strain missing) and model with values substituted into parameters (RheoModel), return a new dataset based on the model. If data is type of stress_only, then creep modulus (:J) is used; if data type is strain_only relaxation modulus (:G). A complete RheoTimeData of type strain_and_stress is returned. diff_method sets finite difference for calculating the derivative used in the hereditary integral and can be either backwards difference ("BD") or central difference ("CD").

RHEOS.modelstepfit — Function

modelstepfit(data::RheoTimeData, model::RheoModelClass, modloading::Union{LoadingType,Integer}; step=nothing, p0::Union{NamedTuple,Nothing} = nothing, lo::Union{NamedTuple,Nothing} = nothing, hi::Union{NamedTuple,Nothing} = nothing, verbose::Bool = false, rel_tol = 1e-4, weights::Union{Vector{Integer},Nothing} = nothing)

Same as modelfit except assumes a step loading. If this assumption is appropriate for the data then fitting can be sped up greatly by use of this function. If modloading is strain_imposed, relaxation modulus is used, then the element in the middle of the strain is assumed to be the amplitude of the step. If modloading is stress_imposed, the creep modulus is used, then the middle element of the stress is assumed to be the amplitude of the step. Alternatively, it is possible to define the value of the step by defining the optional step parameter.

Arguments

data: RheoTimeData struct containing all data

model: RheoModelClass containing moduli and parameters tuples

modloading: strain_imposed for relaxation modulus, stress_imposed for creep modulus

step: Optional amplitude for step

p0: Named tuple of initial parameters to use in fit (uses 0.5 for all parameters if none given)

lo: Named tuple of lower bounds for parameters

hi: Named tuple of upper bounds for parameters

verbose: If true, prints parameters on each optimisation iteration

rel_tol: Relative tolerance of optimization, see NLOpt docs for more details

weights: Vector of indices weighted according to importance, can be generated by indexweight function

RHEOS.modelsteppredict — Function

modelsteppredict(data::RheoTimeData, model::RheoModel; step_on::Real = 0.0)

Same as modelpredict but assumes a step loading with step starting at step_on or closest actual value to that specified. If the loading data is variable, the magnitude in the middle of the array is used.

RHEOS.dynamicmodelfit — Function

dynamicmodelfit(data::RheoFreqData, model::RheoModelClass; p0::Union{NamedTuple,Nothing} = nothing, lo::Union{NamedTuple,Nothing} = nothing, hi::Union{NamedTuple,Nothing} = nothing, verbose::Bool = false, rel_tol = 1e-4) where T<:Real

Fits model to the frequency/loss+storage moduli data.

All arguments are as described below. As this fitting procedure is fitting two functions simultaneously (the storage and loss moduli), if left untransformed the fit would tend to favour the modulus which is larger in magnitude and not fit the other modulus well. To avoid this, RHEOS offers a number of data transforms which can be used by changing weights argument.

Arguments

data: RheoFreqData struct containing all data

model: RheoModelClass containing moduli and symbols of parameters

p0: Initial parameters to use in fit (uses 0.5 for all parameters if none given)

lo: Lower bounds for parameters

hi: Upper bounds for parameters

verbose: If true, prints parameters on each optimisation iteration

rel_tol: Relative tolerance of optimization, see NLOpt docs for more details

weights: Weighting mode for storage and loss modulus ("none", "mean", "log", "local" or manually specified)

RHEOS.dynamicmodelpredict — Function

dynamicmodelpredict(data::RheoFreqData, model::RheoModel)

Given dynamic rheology data with only frequency and model where parameters have been substituted. Returns another RheoFreqData instance with the predicted Gp and Gpp based on the frequencies and model given as arguments.

RHEOS.freeze_params — Function

freeze_params(m::RheoModelClass, nt0::NamedTuple)

Return a new RheoModelClass with some of the parameters frozen to specific values

Arguments

m: original RheoModelClass

nt0: named tuple with values for each parameter to freeze

Example

julia> SLS2_mod = freeze_params( SLS2, (G₀=2,η₂=3.5))
[...]

julia> SLS2.G(1,[2,1,2,3,3.5])
3.8796492

julia> SLS2_mod.G(1,[1,2,3])
3.8796492

Data Generation Functions

RHEOS.timeline — Function

timeline(r)

Generate RheoTimeData struct with only the time data based on the range provided.

Example

timeline(0:0.1:10)   --> 0.  0.1   0.2  ...  9.9  10.

timeline(;t_start::Real=0., t_end::Real, step::Real=(t_end - t_start)/250.)

Generate RheoTimeData struct with only the time data.

Arguments

t_start: Starting time, typically 0

t_end: End time

step: Time between sample. Default set in order to provide ≈ 250 time points.

RHEOS.strainfunction — Function

strainfunction(data::RheoTimeData, f::T) where T<:Function

Accepts a RheoTimeData and outputs a new RheoTimeData with a strain imposed. The strain signal is determined by the function provided, which should take time as its only argument. The original data's time signal is used.

Normally used with a RheoTimeData generated using the timeline function.

RHEOS.stressfunction — Function

stressfunction(data::RheoTimeData, f::T) where T<:Function

Accepts a RheoTimeData and outputs a new RheoTimeData with a stress imposed. The stress signal is determined by the function provided, which should take time as its only argument. The original data's time signal is used.

Normally used with a RheoTimeData generated using the timeline function.

RHEOS.hstep — Function

hstep(t; offset=0., amp=1.)

Step generation function for use with stressfunction or strainfunction. offset keyword arguent determines start of step. amp argument determines amplitude (height) of step.

RHEOS.ramp — Function

ramp(t; offset=0., gradient=1.)

Ramp signal generation function for use with stressfunction or strainfunction. offset keyword argument determines start of ramp. gradient argument determines the linear gradient of the ramp.

RHEOS.stairs — Function

stairs(t; offset=0., amp=1., width=1.)

Stairs signal generation function for use with stressfunction or strainfunction. Equivalent to additional steps being added every width seconds. offset keyword argument determines start of stairs signal. amp argument determines the height of each additional step.

RHEOS.square — Function

square(t; offset=0., amp=1., period=1., width=0.5*period)

Square signal generation function for use with stressfunction or strainfunction. offset keyword argument determines start of square signal. amp argument determines the height of each square pulse. period determines the period of one off/on section of the square wave signal. width determines the width of each square pulse.

RHEOS.sawtooth — Function

sawtooth(t; offset=0., amp=1., period=1.)

Sawtooth signal generation function for use with stressfunction or strainfunction. offset keyword argument determines start of sawtooth signal. amp argument determines the height of each sawtooth pulse. period determines the period of the sawtooth wave signal.

RHEOS.triangle — Function

triangle(t; offset=0., amp=1., period=1.)

Triangle signal generation function for use with stressfunction or strainfunction. offset keyword argument determines start of triangle signal. amp argument determines the height of each triangle pulse. period determines the period of the triangle wave signal. width determines the width of the triangles.

RHEOS.frequencyspec — Function

frequencyspec(r::R; logscale=true)

Generate RheoFreqData struct with only the frequency data distributed on a linear or log scale

Arguments

r: range, e.g. 1:0.1:100

logscale: if true, set the range in logscale, e.g. -2:0.1:20 –> 10 values per decade between 10^-2 and 10^2

frequencyspec(;ω_start::Real, ω_end::Real, logstep::Real= log10(ω_end/ω_start)/10.)

Generate RheoFreqData struct with only the frequency data distributed on a log scale

Arguments

ω_start: min frequency

ω_end: max frequency

logstep: step between frequencies in log scale, e.g. 0.1 –> 10 values per decade.

Data IO

RHEOS.importcsv — Function

importcsv(filepath::String; t_col::IntOrNone = nothing, σ_col::IntOrNone = nothing, ϵ_col::IntOrNone = nothing, ω_col::IntOrNone = nothing, Gp_col::IntOrNone = nothing, Gpp_col::IntOrNone = nothing, delimiter=',')

Load data from a CSV file (two/three columns, comma seperated by default but delimiter can be specified in the delimiter keyword argument). Arguments must be identified by providing the number of the column in which they are contained.

Can be used to construct either a RheoTimeData instance or a RheoFreqData instance. Function detects whether time or frequency has been included and proceeds accordingly. For oscillatory data, all three columns (Gp, Gpp, Frequency) must be provided. For regular viscoelastic data only time, or time-stress, or time-strain or time-stress-strain data can be provided.

RHEOS.exportcsv — Function

exportcsv(self::Union{RheoTimeData, RheoFreqData}, filedir::String; delimiter=',', colorder=nothing)

Export RheoTimeData or RheoFreqData type to csv format. May be useful for plotting/analysis in other software. By default, full time data will be exported with columns ordered as (t, σ, ϵ). Partial time data will be ordered as either (t, σ) or (t, ϵ). Full frequency data will be ordered as (ω, Gp, Gpp). The order of columns can be customised by passing a NamedTuple to the colorder arguments. For example (σ = 1, t = 3, ϵ = 2) would export the columns in the order (σ, ϵ, t). As with importcsv, the delimiter can be set by keyword argument.