API

API

Sampling and Filtering Functions

RHEOS.resample — Function.

resample(self::RheoTimeData, elperiods::Union{Vector{K}, K}; time_boundaries::Vector{T}= [-1])

Resample data with new sample rate(s).

Resample can downsample or upsample data. If the number of elperiods is negative it is going to reduce the number of samples, viceversa if it is positive. If time boundaries are not specified, resampling is applied to the whole set of data. If number of elements per period (elperiods) is 1 or -1 it returns the original RheoTimeData, whilst 0 is not accepted as a valid argument for elperiods.

source

RHEOS.cutting — Function.

cutting(self::RheoTimeData, time_on::T1, time_off::T2) where {T1<:Number, T2<:Number}

Remove the data outside a specified time interval.

By specifing a time interval (timeon, timeoff), a new RheoTimeData is returned without the data laying outside time interval.

source

RHEOS.smooth — Function.

smooth(self::RheoTimeData, τ::Real; pad::String="reflect")

Smooth data using a Gaussian Kernel to time scale τ (approximately half power).

Smooths both σ and ϵ. Essentially a low pass filter with frequencies of 1/τ being cut to approximately half power. For other pad types available see ImageFiltering documentation.

source

RHEOS.extract — Function.

extract(self::Union{RheoTimeData,RheoFreqData}, type::Union{TimeDataType,FreqDataType,Integer})

Extract specific fields form RheoTimeData or RheoFreqData.

Extract can copy one or more fields from a given RheoXData variable into a new RheoXData one. The fields that are copied are identified by the specified type of data. If self is a RheoTimeData, the type that can be extracted is timeonly (or 0), stressonly (or 1), strainonly (or 2). Note that strainandstress (or 3) is not allowed. If self is a RheoFreqData, the type that can be extracted is freconly (or 0).

source

Fitting and Predicting Functions

RHEOS.modelfit — Function.

modelfit(data::RheoTimeData, model::RheoModelClass, modloading::Symbol; p0::Union{NamedTuple,Tuple} = (), lo::Union{NamedTuple,Tuple} = (), hi::Union{NamedTuple,Tuple} = (), verbose::Bool = false, rel_tol = 1e-4, diff_method="BD")

Fit RheologyData struct to model and return a fitted model as a RheologyModel object.

Arguments

data: RheoTimeData struct containing all data

model: RheoModelClass containing moduli functions and named tuple parameters

modloading: strainimposed or 1, stressimposed or 2

p0: Initial parameters to use in fit (uses 0.5 for all parameters if not defined)

lo: Lower bounds for parameters

hi: Upper bounds for parameters

verbose: If true, prints parameters on each optimisation iteration

rel_tol: Relative tolerance of optimization, see NLOpt docs for more details

diff_method: Set finite difference formula to use for derivative, currently "BD" or "CD"

source

RHEOS.modelpredict — Function.

modelpredict(data::RheoTimeData,model::RheoModel; diff_method="BD")

Given an incomplete data set (only either stress or strain missing) and model with values substituted into parameters (RheoModel),return a new dataset based on the model. If data is type of stressonly, then creep modulus (:J) is used; if data type is strainonly relaxation modulus (:G). A complete RheoTimeDatadata of type "strainandstress" is returned. 'diff_method' sets finite difference for calculating the derivative used in the hereditary integral and can be either backwards difference ("BD") or central difference ("CD").

source

RHEOS.modelstepfit — Function.

modelstepfit(data::RheoTimeData, model::RheoModelClass, modloading::Union{LoadingType,Integer}; step=nothing, p0::Union{NamedTuple,Tuple} = (), lo::Union{NamedTuple,Tuple} = (), hi::Union{NamedTuple,Tuple} = (), verbose::Bool = false, rel_tol = 1e-4) where T<:Real

Same as 'modelfit' except assumes a step loading. If this assumption is appropriate for the data then fitting can be sped up greatly by use of this function. If modloading is strainimposed, relaxation modulus is used, then the element in the middle of the strain is assumed to be the amplitude of the step. If modloading is stressimposed, the creep modulus is used, then the middle element of the stress is assumed to be the amplitude of the step. Alternatively, it is possible to define the value of the step by defining the optional "step" parameter.

Arguments

data: RheoTimeData struct containing all data

model: RheoModelClass containing moduli and parameters tuples

modloading: strainimposed for relaxation modulus, stressimposed for creep modulus

p0: Named tuple of initial parameters to use in fit (uses 0.5 for all parameters if none given)

lo: Named tuple of lower bounds for parameters

hi: Named tuple of upper bounds for parameters

verbose: If true, prints parameters on each optimisation iteration

rel_tol: Relative tolerance of optimization, see NLOpt docs for more details

diff_method: Set finite difference formula to use for derivative, currently "BD" or "CD"

source

RHEOS.modelsteppredict — Function.

modelsteppredict(data::RheoTimeData, model::RheoModel; step_on::Real = 0.0, diff_method = "BD")

Same as modelpredict but assumes a step loading with step starting at 'step_on'. Singularities are bypassed by adding 1 to the index of the singular element.

source

RHEOS.dynamicmodelfit — Function.

dynamicmodelfit(data::RheoFreqData, model::RheoModelClass; p0::Union{NamedTuple,Tuple} = (), lo::Union{NamedTuple,Tuple} = (), hi::Union{NamedTuple,Tuple} = (), verbose::Bool = false, rel_tol = 1e-4) where T<:Real

Fits model to the frequency/loss+storage moduli data.

All arguments are as described below. As this fitting procedure is fitting two functions simultaneously (the storage and loss moduli), if left untransformed the fit would tend to favour the modulus which is larger in magnitude and not fit the other modulus well. To avoid this, RHEOS offers a number of data transforms which can be used by changing "weights" argument.

Arguments

data: RheoFreqData struct containing all data

model: RheoModelClass containing moduli and symbols of parameters

p0: Initial parameters to use in fit (uses 0.5 for all parameters if none given)

lo: Lower bounds for parameters

hi: Upper bounds for parameters

verbose: If true, prints parameters on each optimisation iteration

rel_tol: Relative tolerance of optimization, see NLOpt docs for more details

weights: Weighting mode for storage and loss modulus (linear, log, global)

source

RHEOS.dynamicmodelpredict — Function.

dynamicmodelpredict(data::RheoFreqData, model::RheoModel)

Given dynamic rheology data with only frequency and model where parameters have been substituted. Returns another RheoFreqData instance with the predicted Gp and Gpp based on the frequencies and model given as arguments.

source

Data Generation Functions

RHEOS.timeline — Function.

timeline(;t_start::Real=0., t_end::Real=10., step::Real=(t_end - t_start)/250.)

Generate RheoTimeData struct with only the time data.

Arguments

t_start: Starting time, typically 0

t_end: End time

step: Time between sample

source

RHEOS.strainfunction — Function.

strainfunction(data::RheoTimeData, f::T) where T<:Function

Accepts a RheoTimeData and outputs a new RheoTimeData with a strain imposed. The strain signal is determined by the function provided, which should take time as its only argument. The original data's time signal is used.

Normally used with a RheoTimeData generated using the timeline function.

source

RHEOS.stressfunction — Function.

stressfunction(data::RheoTimeData, f::T) where T<:Function

Accepts a RheoTimeData and outputs a new RheoTimeData with a stress imposed. The stress signal is determined by the function provided, which should take time as its only argument. The original data's time signal is used.

Normally used with a RheoTimeData generated using the timeline function.

source

RHEOS.hstep — Function.

hstep(t; offset=0., amp=1.)

Step generation function for use with stressfunction or strainfunction. offset keyword arguent determines start of step. amp argument determines amplitude (height) of step.

source

RHEOS.ramp — Function.

ramp(t; offset=0., gradient=1.)

Ramp signal generation function for use with stressfunction or strainfunction. offset keyword arguent determines start of ramp. gradient argument determines the linear gradient of the ramp.

source

RHEOS.stairs — Function.

stairs(t; offset=0., amp=1., width=1.)

Stairs signal generation function for use with stressfunction or strainfunction. Equivalent to additional steps being added every width seconds. offset keyword arguent determines start of stairs signal. amp argument determines the height of each additional step.

source

RHEOS.square — Function.

square(t; offset=0., amp=1., period=1., width=0.5*period)

Square signal generation function for use with stressfunction or strainfunction. offset keyword arguent determines start of square signal. amp argument determines the height of each square pulse. period determines the period of one off/on section of the square wave signal. width determines the width of each square pulse.

source

RHEOS.sawtooth — Function.

sawtooth(t; offset=0., amp=1., period=1.)

Sawtooth signal generation function for use with stressfunction or strainfunction. offset keyword arguent determines start of sawtooth signal. amp argument determines the height of each sawtooth pulse. period determines the period of the sawtooth wave signal.

source

RHEOS.triangle — Function.

triangle(t; offset=0., amp=1., period=1.)

Triangle signal generation function for use with stressfunction or strainfunction. offset keyword arguent determines start of triangle signal. amp argument determines the height of each triangle pulse. period determines the period of the triangle wave signal. width determines the width of the triangles.

source

RHEOS.frequencyspec — Function.

frequencyspec(;ω_start::Real=1.0e-2, ω_end::Real=1.0e2, step::Real=(ω_end-ω_start)/1.0e5)

Generate RheoFreqData struct with only the frequency data.

Arguments

ω_start: Starting time, typically 0

ω_end: End time

step: Step between frequencies

source

Data IO

RHEOS.importcsv — Function.

importcsv(filepath::String; t_col::IntOrNone= nothing, σ_col::IntOrNone= nothing, ϵ_col::IntOrNone=nothing, ω_col::IntOrNone= nothing, Gp_col::IntOrNone = nothing, Gpp_col::IntOrNone = nothing, delimiter=',')

Load data from a CSV file (two/three columns, comma seperated by default but delimiter can be specified in the delimiter keyword argument). Arguments must be identified by providing the number of the column in which they are contained.

Can be used to construct either a RheoTimeData instance or a RheoFreqData instance. Function detects whether "time" or "frequency" has been included and proceeds accordingly. For oscillatory data, all three columns (Gp, Gpp, Frequency) must be provided. For regular viscoelastic data only time, or time-stress, or time-strain or time-stress-strain data can be provided.

source

RHEOS.exportcsv — Function.

exportcsv(self::Union{RheoTimeData, RheoFreqData}, filedir::String; delimiter=',', colorder=nothing)

Export RheoTimeData or RheoFreqData type to csv format. May be useful for plotting/analysis in other software. By default, full time data will be exported with columns ordered as (σ, ϵ, t). Partial time data will be ordered as either (σ, t) or (ϵ, t). Full frequency data will be ordered as (Gp, Gpp, ω). The order of columns can be customised by passing a NamedTuple to the colorder arguments. For example (σ = 2, t = 1, ϵ = 3) would export the columns in the order (t, σ, ϵ). As with importcsv, the delimiter can be set by keyword argument.

source